requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
groq==0.3.0
flask_cors==4.0.1
python-dotenv==1.0.0